+++ pycep/formatter.py
@@ -13,6 +13,16 @@ def add_newline(input_item: str) -> str:
     return f"{input_item}\n"
 
 
+def add_whitespace(input_string):
+    url_safe_string = ""
+    for char in input_string:
+        if char != " ":
+            url_safe_string += char
+        else:
+            url_safe_string += "%20"
+    return url_safe_string
+
+
 def strip_unsafe_file_names(string_data: str) -> str:
     new_string = ""
     for character in string_data: